嘘...

+++ src/minecraft/cn/stars/reversal/font/MFont.java
@@ -1,13 +1,13 @@
 package cn.stars.reversal.font;
 
 public abstract class MFont {
-    public abstract float drawString(String text, double x, double y, int color, boolean dropShadow);
+    public abstract int drawString(String text, double x, double y, int color, boolean dropShadow);
 
-    public abstract float drawString(final String text, final double x, final double y, final int color);
+    public abstract int drawString(final String text, final double x, final double y, final int color);
 
-    public abstract float drawStringWithShadow(final String text, final double x, final double y, final int color);
+    public abstract int drawStringWithShadow(final String text, final double x, final double y, final int color);
 
-    public abstract float width(String text);
+    public abstract int width(String text);
 
     public float getWidth(String text) { return width(text); }
